+++ docs/changelog.rst
@@ -7,6 +7,8 @@ Changelog
 Development:
   * Modify get status integration test schema to support new
     ``indexingEnabled`` field
+  * Add/modify integration tests for new smart contract secret restrictions
+  * Deprecate ETC Testnet (morden)
 
 4.2.0
 -----